requirements.txt
numpy==1.25.2
pandas==2.1.0
Pillow==9.5.0
Pillow==10.0.0
streamlit==1.26.0
streamlit_option_menu==0.3.6
